refactores favorite list
